Use Numeric rather than Integer for the validation

+++ lib/durationizer.rb
@@ -37,7 +37,7 @@ module Durationizer
                                in: VALID_UNIT_TYPES,
                                message: 'has to be a valid unit, eg seconds, minutes, hours, etc'
         validate do |model|
-          model.errors.add(:base, 'Frequency has to be a valid number') unless model.send(column).is_a?(Integer)
+          model.errors.add(:base, 'Frequency has to be a valid number') unless model.send(column).is_a?(Numeric)
         end
       end
     end
